Conflicts:
	src/ATO/ATO_Solver.cpp

+++ src/utility/CounterMonitor.cpp
@@ -12,7 +12,7 @@ CounterMonitor::CounterMonitor () {
 }
 
 string CounterMonitor::getStringValue (const monitored_type& val) {
-  return std::to_string(val.value());
+  return std::to_string(static_cast<long long>(val.value()));
 }
 
 }
